refactor: HandTracker.tsx 파일 구조 세분화
