+++ src/features/handtracking/handState.ts
